Move dictionary related functions to dictioanry package
